requirements.txt
streamlit
PyPDF2
google-generativeai
python-dotenv
sentence-transformers
faiss-cpu
numpy